Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -2,10 +2,10 @@
 
 int main(){
     int a,b,c,sum;
-    float avg;
+    double avg;
     scanf("%d %d %d",&a,&b,&c);
     sum = a+b+c;
     avg = sum/3;
-    printf("Average: %.2f",avg);
+    printf("Average: %.2lf",avg);
     return 0;
 }
